Colour Palettes

List the available colour palettes.

colour_palettes(colours = NULL)

color_palettes(colours = NULL)

Arguments

colours
vector of source colour palettes to return, one or many of "viridis","rcolorbrewer","grdevices","colorspace","colorramp". NULL will reutrn all palettes.

Details

The palettes avaialble in colourvalues have been derived from those avaialble in the libraries

viridis

RColorBrewer

grDevices

colorspaces

colorRamp

Examples

colour_palettes()

#>  [1] "viridis"        "cividis"        "magma"          "inferno"       
#>  [5] "plasma"         "ylorrd"         "ylorbr"         "ylgnbu"        
#>  [9] "ylgn"           "reds"           "rdpu"           "purples"       
#> [13] "purd"           "pubugn"         "pubu"           "orrd"          
#> [17] "oranges"        "greys"          "greens"         "gnbu"          
#> [21] "bupu"           "bugn"           "blues"          "spectral"      
#> [25] "rdylgn"         "rdylbu"         "rdgy"           "rdbu"          
#> [29] "puor"           "prgn"           "piyg"           "brbg"          
#> [33] "terrain"        "topo"           "heat"           "cm"            
#> [37] "rainbow"        "terrain_hcl"    "heat_hcl"       "sequential_hcl"
#> [41] "rainbow_hcl"    "diverge_hcl"    "diverge_hsv"    "ygobb"         
#> [45] "matlab_like2"   "matlab_like"    "magenta2green"  "cyan2yellow"   
#> [49] "blue2yellow"    "green2red"      "blue2green"     "blue2red"      

colour_palettes( "viridis" )

#> [1] "viridis" "cividis" "magma"   "inferno" "plasma" 

colour_palettes( colours = c("rcolorbrewer","grdevices") )

#>  [1] "ylorrd"   "ylorbr"   "ylgnbu"   "ylgn"     "reds"     "rdpu"    
#>  [7] "purples"  "purd"     "pubugn"   "pubu"     "orrd"     "oranges" 
#> [13] "greys"    "greens"   "gnbu"     "bupu"     "bugn"     "blues"   
#> [19] "spectral" "rdylgn"   "rdylbu"   "rdgy"     "rdbu"     "puor"    
#> [25] "prgn"     "piyg"     "brbg"     "terrain"  "topo"     "heat"    
#> [31] "cm"       "rainbow"